Some really hacky code
Currently Skype Web appears to invalidate a login after having been logged
on for too long. The exact time is not known but there does not appear to
exist a native solution (eg one offered by Skype Web API) as the web
client will crash after being connected for too long
(fullExperience.min.js will stop responding which I assume is caused by
the invalidated credentials).

In the near future a system for alerting developers of this invalidation
as well as better handling of it will be implemented, but until then...
you have this