Major internal changes
This commit fixes quite a few major internal issues. Let's get started.

First of all, if Skype4J gets a richtext message with some tags it doesn't
recgonize, it'll throw an exception because of invalid handling. That's
been fixed... sort of

If your account is already logged in or was recently logged in, you may
already be assigned to a cloud. If so, Skype4J wasn't able to log you back
in and would throw a IOException due to 403. That's fixed now.

Skype.java didn't have a public login() method. That's been fixed.

Adding yourself into a group would result in an error because the event
was mis-handled. That's been fixed

NestedSkypeException didn't properly add the suppressed exception.
Code was organized in the internal packages
Some other tweaks here and there